Building upgrade

+++ src/main/java/com/denamed/TestWMS/services/BuildingService.java
@@ -2,6 +2,7 @@ package com.denamed.TestWMS.services;
 
 import com.denamed.TestWMS.entities.Building;
 import com.denamed.TestWMS.repositories.BuildingRepository;
+import com.denamed.TestWMS.repositories.ModuleRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -11,26 +12,32 @@ import java.util.List;
 public class BuildingService {
 
     private final BuildingRepository buildingRepository;
+    private final ModuleRepository moduleRepository;
 
     @Autowired
-    public BuildingService(BuildingRepository buildingRepository) {
+    public BuildingService(BuildingRepository buildingRepository, ModuleRepository moduleRepository) {
         this.buildingRepository = buildingRepository;
+        this.moduleRepository = moduleRepository;
     }
 
     public List<Building> findAll(){
         return buildingRepository.findAll();
     }
 
-    public Building addBuilding(Building building) throws Exception{
+    public Building createBuilding(Building building) throws Exception{
         if(buildingRepository.existsById(building.getBuildId())) {
-            throw new Exception("<style>.alert {color: #FF3300;}</style><h2 class=\"alert\">Building with id " + building.getBuildId() + " already exists</h2>");
+            throw new Exception("<style>.alert {color: #FF3300;}</style><h2 class=\"alert\">Building " + building.getBuildId() + " already exists!</h2>");
         } else {
             return buildingRepository.save(building);
         }
     }
 
-    public void deleteBuilding(int buildId){
-        buildingRepository.deleteById(buildId);
+    public void deleteBuilding(int buildId) throws Exception{
+        if (moduleRepository.existsByBuildId(buildId)) {
+            throw new Exception("<style>.alert {color: #FF3300;}</style><h2 class=\"alert\">Building " + buildId + " have the modules!</h2>");
+        } else {
+            buildingRepository.deleteById(buildId);
+        }
     }
 
 }
